gestion des tables users et utilisateurs ainsi que les models
